Update semantic3d.py
If this modification is not made, it means that the labels of test files can't be read. In that case, the code block `if (gt_labels > 0).any():` in `torch/pipelines/semantic_segmentation.py` will be always ignored when executing `pipeline.run_test()`, which causes the skipping of the following metric calculation.